+++ testing/premake5.lua
@@ -62,7 +62,7 @@ project "init"
 		( "rm -rf ./obj/" ),
 		( "rm -rf ./bin/" ),
 
-		("@echo \"Clean testing will have an error - don't worry, its fine (it seems to be a bug in the genation from premake5)\""),
+		("@echo \"Clean testing may have an error - don't worry, its fine (it seems to be a bug in the genation from premake5)\""),
 		("@echo"),
 	}